A hack patch solves the issues we have with pgfkeys and keyval when a key has several value with delimited by { } src.: http://tex.stackexchange.com/a/322238/66964
xkeyval is *still* affected by this issue and seems buggy, the
previously solutions implemented with keyval or pgfkeys are now working.

Performing this simple test is now working thanks to this hack. Removing
it makes the compilation crash. Babel is still loaded and loading it
does not clash when we use several values for one key. Btw, babel was
not the only package affected by this issue, for a list see the question
that has been updated accordingly:
http://tex.stackexchange.com/q/322217/66964

\makeatletter
\let\oldin@\in@
\def\zzz#1#2#{\oldin@{#2}{#1}\zzzb}
\def\in@#1#2{\zzz{#2}#1{}\@nil}
\def\zzzb#1\@nil{}
\makeatother

\RequirePackage{filecontents}
\begin{filecontents*}{testtemplate.cls}
\ProvidesClass{testtemplate}
\LoadClass[twoside]{report}
\RequirePackage[english]{babel}
\end{filecontents*}
\documentclass[someOtherKey=hello
world,languages={dutch,french}]{testtemplate}
\begin{document}
    Hello world
\end{document}